requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyPDF2==3.0.1
PyMuPDF==1.25.0
Pillow==10.1.0
pytest==8.3.5
pytest-django==4.11.1
python-dotenv==1.0.1
gunicorn==21.2.0
whitenoise==6.6.0
coreapi==2.3.3
WeasyPrint==57.1
django-filter==23.1
fpdf==1.7.2
reportlab==4.0.4
psycopg2-binary==2.9.9
dj-database-url==2.1.0
supabase==2.7.4
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
djangorestframework-simplejwt==5.3.0